add interaction test
